week6-week7/crAPI/services/chatbot/requirements.txt
langchain==1.0.3
langchain-community==0.4.1
langchain-core==1.0.2
langchain-mongodb==0.7.1
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langchain-mcp-adapters==0.1.11
langgraph==1.0.2
chromadb==1.3.0
langchain-chroma==1.0.0
openai==2.6.1
fastmcp==2.13.0.2
markdown-it-py==3.0.0
pymongo==4.15.3
python-dotenv==1.2.1
unstructured==0.18.15
numpy==2.2.6
Quart==0.20.0
quart-cors==0.8.0
motor==3.7.1
faiss-cpu==1.12.0
psycopg2-binary==2.9.11
uvicorn==0.38.0
langchain-aws
langchain-google-vertexai
langchain-anthropic
langchain-groq
langchain-mistralai
langchain-cohere